+++ QSB/Player/PlayerInfoParts/DreamWorld.cs
@@ -1,4 +1,6 @@
-﻿using QSB.ItemSync.WorldObjects.Items;
+﻿using QSB.EchoesOfTheEye.DreamLantern;
+using QSB.EchoesOfTheEye.DreamLantern.WorldObjects;
+using QSB.ItemSync.WorldObjects.Items;
 
 namespace QSB.Player;
