+++ QSB/Utility/DebugActions.cs
@@ -1,4 +1,6 @@
 ﻿using OWML.Common;
+using QSB.EchoesOfTheEye.DreamLantern;
+using QSB.EchoesOfTheEye.DreamLantern.WorldObjects;
 using QSB.ItemSync.WorldObjects.Items;
 using QSB.Messaging;
 using QSB.Player;
